test: add validator tests
